backend/requirements.txt
# Web framework
fastapi==0.109.0

# Web server
uvicorn==0.27.0
websockets>=12.0

# Google Gemini AI (google.genai SDK)
google-genai>=0.6.0

# agent orchestration
langchain>=0.1.0
langgraph>=0.0.20
langchain-core>=0.1.0

# PDF Processing & Text Extraction
pdfplumber>=0.10.0
requests>=2.31.0

# Vector Embeddings & Semantic Search
sentence-transformers>=2.2.0
chromadb>=0.4.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0

# Search Tools
ddgs>=0.1.0  
feedparser>=6.0.11  
arxiv>=2.1.0  
httpx==0.27.0  

# Environment variables
python-dotenv==1.0.0
